Basic Bluetooth Discovery Implemented
This is a large commit including several modification.

1. The LogIn and SignUp activities have been cleaned up, organized, and optimized. The PHP code relating to this is fixed as well and should avoid SQL injection while allowing for apostrophes and other inputs to still work.

The email still follows a specified syntax to avoid invalid emails, and any spaces are stripped from the input. Emails are case sensitive when stored, however, that is not considered when signing up.

The username input has been overhauled and only allows a-z, A-Z, 0-9, spaces, -. _. and '. The input is limited to 32 characters and should allow for quite a lot of versatility while still avoid dangerous input as well as characters I may use for filtering. Usernames are also case sensitive in the database so they can be displayed the way a user may want it to be displayed. They are not case sensitive when comparing inputs though.

Passwords are no longer cleaned since they are hashed right away so they can't do any harm. This should help avoid any issues that may occur when un-hashing values.

2. A device activity has been added which displays a list of all feeders owned by the user. I'm hoping to eventually add a bit more style to this page, however, it works for now. Eventually the cards should be clickable and will take to you a page with all of the information and settings regarding that specific device. In the bottom right of this activity is an add button. Clicking this bring the user to the Bluetooth activity which will be described next.

3. A Bluetooth discovery activity has been added and is in it's simplest form right now. It consists of a scroll view and a text view to display any devices found. I think I will eventually change this over to a card view to better view everything, but I wanted to upload this before any more changes occur. I have also added a filter to only display devices following the naming scheme "DSD ...." I would ideally configure all feeders with this prefix so only feeders will be displayed in the device list.

4. The next thing to do is modify the way the discovered devices are displayed and make them clickable. When clicked, it should ask the user to enter the password located on the feeder itself. This will prevent unwanted connections while still allowing multiple users to connect to and monitor a single device if desired. If the password is correct, the device should connect to the feeder and request the user to input their Wi-Fi SSID and password. I was thinking about making this a global setting, however, there is the possibility that a user would have feeders on separate connections.

After the user inputs the Wi-Fi information, the device should check to see if a connection can be established or not. If not, it should ask the user to input their information again. If an error still occurs, I could possibly store the values in the users phone until a connection can be established. This would be an issue regarding Bluetooth connections though.

If the connection is established, it should upload the device information to the database and add the user as its owner. Once the connection is established, it should disconnect from Bluetooth and all communication should occur over the internet.

This is the point where I want to get to soon, so hopefully my next commit will be working toward that goal. There is plenty left to do after reaching this point, however, once this is set up, I can finally start working with the device. Everything up until then is just to establish a connection.