Bluetooth Devices Basic Card View
Switched the Bluetooth device list to a card view just like the feeder list.

Changed some naming schemes to allow for easier understanding.
    - Still need to adjust some naming conventions since they aren't consistent, but I'm not worried about that right now.

Fixed some issues in the code as well as added the ability to show paired devices for debugging purposes. This is commented out and needs to be edited to allow for it.

Next step is to implement clicking for cards in the Bluetooth device list.