Basic BT onclick listener implemented
Added a basic onclick listener to the Bluetooth list RecyclerView. This allows users to click on the Bluetooth device they want to connect to and it will start a new activity to input login details.

The layout of the new activity is set up and an untested threaded class used to try to connect to the Bluetooth device is added. Along with this, the class for Bluetooth communication if connected is added but has no code.

Some code was cleaned up, some was removed, and overall, things just changed. Ran into a few issues trying to decide who would have the Bluetooth object in the end but decided it would be best for the Wi-Fi activity to have it since it will need to send data for Wi-Fi connectivity.

Next step is to try connecting to the Bluetooth device and possibly sending data.