Not Working Backup
Just a quick backup before modifying a lot of code.
While I thought I had tested it before, it seems I need to switch the code over to work with BLE. I can't connect to the Bluetooth module right now and is most likely due to it being a BLE device and not a typical Bluetooth device. Guess we will see if a GATT server works...