Finally have Bluetooth connection!
This commit is the first one that establishes an actual Bluetooth connection. After modifying the code to work with BLE devices, I had to create a service to handle the GATT server. This initially didn't work, however, I finally realized we need to define the service in the Manifest so the program knows what it is.

From here, I can start learning how to communicate and transfer data. That means the next step is to transfer the Wi-Fi credentials.