Access to HM-10 services and characteristics
This commit is a quick backup where I finally got to display a list of services and characteristics of the HM-10 device. I initially couldn't identify what they were, but I found a document that sumerises them. Importantly, the final service is used to transmit data so that is what I will be working with.

It's important to note that the current setup is temporary. I was just trying to see information so I could find what service I need to work with.

Next, I need to label these services and start writing data and reading data from service 3.