Successful Reading and Writing Data
This commit allows for proper reading and writing of data. It should be noted though that reading from the specified characteristic doesn't seem to change even after being written to. Either I am messing up somewhere or the characteristic value doesn't change since it is used for serial communication. Along with this, the actual value being read from the characteristic is the mac address of the device.

Either way, I can properly send data to the Arduino Serial port and can receive data since writing it from the Arduino triggers a characteristic change broadcast.

Next, I need finalize the layout and remove the debugging outputs. Then, I can send the user input and use it to try to connect to the internet.