Updated write function to allow for longer inputs
This commit changes the write function so it can handle inputs that are longer than 20 bytes. This is a limitation of the HM-10 characteristic so I need to only send a maximum of 20 bytes at a time. This was accomplished by having a function that writes data to a queue and a function that removes up to 20 values from that queue and sends the data. Importantly, this won't send any more data than is necessary (like zeros).