Ability to add feeders to database
It's absolutely scuffed, but I am finally able to add devices to the database. That means I have established a Bluetooth connection, sent Wi-Fi credentials over it, checked if they work or not, and if they do, send the device information to the database. There are a lot of problems with the way it is set up right now though.

First, the Bluetooth connection is inconsistent. I have gotten it to stay connected on the odd chance that it disconnects, however, I still don't have a good way of making sure it gets disconnected after it's done. This is most likely due to something not closing properly, but I can't find it for the life of me. Because of this, I may have to add a paring button to the Arduino circuit that allows me to power cycle the Bluetooth device since that's to only way I can find to reset it.

Second, input validation is pretty much out the window at this point. I tried to make sure it was good initially, but it's taking way to much time for something that won't matter since I'm the one using the device. If this was going to be distributed, then I would want to fix it. But until that happens (which it wont), I am going to try to ignore it as much as possible.

Lastly, the code is just shit, especially the connection based section. It was pretty organized at first, but I was struggling to get the connection to close so now I have code in a few places that might overlap, I'm not sure. Either way, it's bad code and I can't be bothered to fix it until I finish the project.

Anyway, with all of that done, I can finally work on adding the main program. It's kind of sad to think that I could have skipped all of this and just hard coded a connection in the first place, but no, I had to make it hard... Either way, the next thing I need to do is add an onClickListener for the FeederList. Luckily, I have already done this for the BluetoothList. After that, I need to add a schedule with food adjustments, a way to rename the device, and a way to remove the device.

The scheduling section is going to be the hardest, but the two other parts should be simple PHP POST requests. All in all, things are starting to get less stressful since I have a path I can see. I'm definitely cutting it close though. Lets hope it works out.