Basic Feeder Interact Implementation
This commit adds a basic GUI for users to interact with their feeder. It does not include a scheduling form yet (that will be next), but it does have the ability to add and remove feeders as well as name them. This commit coincides with the Android commit and will line up with the PHP commit that is to come.

This commit also fixes some errors and cleans up some code.

Next, I need to add a scheduling form that will save the information on the database.