Working Feeding Times List Without Database
This commit adds the Feeder Times List to the Feeder homepage. This turned out way better than I expected, but now I need to find out how I am going to store this information on the database. And once I have done that, I need to be able to interpret the information.

After a bit of testing, I have decided to store the information on the database as JSON. This is because there are library's which make serialization and deserialization very easy, plus databases handle it well. So, once the user opens the feeder, it will retrieve the list of times which will be stored as JSON based on the MAC address of the device and the owner.  Once the list is retrieved, it will populate the screen. The user can then add or remove new feeding times, updating the database with each operation.

Next, I need to actually store the data on the database. This should be pretty simple, but I figured I would commit this working version first.