Working Feeding Times List With Database
This commit builds of the previous one and actually works. The list of times is stored on the database based on mac address. This way, the schedule will be a global setting per feeder, but something like the name of the feeder is based on each user. This allows multiple users to have the same device, give it different names, but see the same feeding schedule. This would be beneficial if more options and settings are added in the future so they can affect who they need to.

At this point, the android app is pretty much done. There are a few issues with repeating messages and I still need to remove some debug messages, but overall, it's working. The next thing I need to do is POST the animal packet to the server on my Arduino and compare it against the schedule using PHP. Then, I need to send back a givefood packet that will either give no food or the specified amount of food in the schedule.